fixed file names, more robust getOppositeDirection
